Refactors module imports and downgrades Tailwind CSS
Updates import paths to use renamed modules for better organization - scheduler and unified search components now have more descriptive names.

Downgrades Tailwind CSS from v4 to v3.4.1 and removes deprecated PostCSS plugin to resolve compatibility issues.
