+++ frontend/package.json
@@ -34,7 +34,6 @@
     "@supabase/auth-helpers-nextjs": "^0.10.0",
     "@supabase/ssr": "^0.6.1",
     "@supabase/supabase-js": "^2.50.0",
-    "@tailwindcss/postcss": "^4.1.10",
     "@tailwindcss/typography": "^0.5.16",
     "@tanstack/react-query": "^5.66.2",
     "@vercel/analytics": "^1.4.5",
@@ -84,7 +83,7 @@
     "postcss": "^8.5.5",
     "prettier": "^3.5.3",
     "prettier-plugin-tailwindcss": "^0.6.10",
-    "tailwindcss": "^4.1.10",
+    "tailwindcss": "^3.4.1",
     "tsx": "^4.20.3",
     "typescript": "^5.8.3"
   }
